Added suffix option and recurse and overwrite flags

+++ ipynb_to_py.py
@@ -3,13 +3,21 @@
 import argparse
 import glob
 import json
+from os.path import exists, join
+import sys
 
-def extract_code(afolder):
-    for infile in glob.glob("./**/*.ipynb", recursive=True):
-        outfile = infile.replace('.ipynb', '.py')
+def extract_code(folder, overwrite, recurse, suffix):
+    glob_string = join(folder, "*.ipynb")
+    if recurse:
+        glob_string = join(folder, "**", "*.ipynb")
+    for infile in glob.glob(glob_string, recursive=recurse):
+        outfile = infile.replace('.ipynb', suffix)
+        if not overwrite and exists(outfile):
+            print(f"Refusing to overwrite: {outfile}")
+            continue
         with open(outfile, 'w') as out_f:
             with open(infile, encoding='utf-8') as in_f:
-                print("Processing:", infile, 'into', outfile)
+                print(f"Processing: {infile} into {outfile}")
                 try:
                     data = json.load(in_f)
                     for acell in data['cells']:
@@ -29,9 +37,22 @@ def extract_code(afolder):
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Extract code from ipynb files')
-    parser.add_argument('--folder', action='store', dest='folder')
+    parser.add_argument('--folder', '-f', dest='folder', action='store',
+                        help='folder to look for ipynb files')
+    parser.add_argument('--overwrite', '-o', dest='overwrite', action='store_true',
+                        help='overwrite files if they exist (default: False)')
+    parser.add_argument('--recurse', '-r', dest='recurse', action='store_true',
+                        help='recurse into subdirectories looking for files (default: False)')
+    parser.add_argument('--suffix', '-s', dest='suffix', action='store',
+                        help='suffix to apply to new files (default: .py)')
+    parser.set_defaults(overwrite=False, recurse=False, suffix='.py')
     args = parser.parse_args()
-    afolder = args.folder
-    extract_code(afolder)
+
+    extract_code(
+        folder=args.folder,
+        overwrite=args.overwrite,
+        recurse=args.recurse,
+        suffix=args.suffix
+    )
